pyproject.toml
[tool.poetry]
name = "sbobinai"
version = "0.0.0"
description = "Simple project template for python projects"
authors = ["Mattia Tantardini <mattia.tantardini@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
tantaroba = "^0.4.0"
openai = "^0.28.0"
openai-whisper = "^20230918"
torch = ">=2.0.0, !=2.0.1"
streamlit = "^1.27.0"
streamlit-audiorec = "^0.1.3"
pyperclip = "^1.8.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"
commitizen = "^2.35.0"
python-semantic-release = "^7.32.1"
pytest = "^7.1.3"

[tool.poetry.group.docs.dependencies]
Sphinx = "^5.3.0"
sphinx-autodoc-typehints = "^1.19.4"
sphinx-rtd-theme = "^1.0.0"
myst-parser = "^0.18.1"

# --- PyTest config -------------------------------------=------ #
[tool.pytest.ini_options]
markers = [
    "functions: mark test for the base functionalities",
    "core: all about the core",
    "db: all about the database",
    "api: all about the api of the system",
]

# --- Commitizen --- #
[tool.commitizen]
name = "cz_conventional_commits"
style = [
    ["qmark", "fg:#ff9d00 bold"],
    ["question", "bold"],
    ["answer", "fg:#ff9d00 bold"],
    ["pointer", "fg:#ff9d00 bold"],
    ["highlighted", "fg:#ff9d00 bold"],
    ["selected", "fg:#cc5454"],
    ["separator", "fg:#cc5454"],
    ["instruction", ""],
    ["text", ""],
    ["disabled", "fg:#858585 italic"]
]
version = "0.0.0"
version_files = [
    "pyproject.toml:version",
    "__version__.py:__version__"
]
tag_format = "v$major.$minor.$patch"
annotated_tag = true
bump_message = "release $current_version -> $new_version"

# --- Semantic release --- #
[tool.semantic_release]
# Reference: https://python-semantic-release.readthedocs.io/en/latest/index.html
version_variable = [
    "pyproject.toml:version"
]
branch = "main"
hvcs = "github"
tag_format = "v{version}"
build_command = ""
upload_to_pypi = false
upload_to_repository = false
commit_subject = "Updated version to v{version}"
